+++ extensions/bitbucket/src/components/pullRequests/searchMyPullRequests.tsx
@@ -96,7 +96,11 @@ export function SearchMyPullRequests() {
               actions={
                 <ActionPanel>
                   <ActionPanel.Section>
-                    <ShowPullRequestDetailAction pr={pr} onDeclined={() => removePullRequest(key)} />
+                    <ShowPullRequestDetailAction
+                      pr={pr}
+                      onDeclined={() => removePullRequest(key)}
+                      onDone={onReviewStateChange}
+                    />
                     <Action.OpenInBrowser
                       title="Open Pull Request in Browser"
                       url={`https://bitbucket.org/${pr.repo.fullName}/pull-requests/${pr.id}`}